fix: some changes to make native permission work
