+++ lib/agent/actions/request_permission/index.js
@@ -1,14 +1,52 @@
 const { EventEmitter } = require('events');
-const common = require('../../../common');
-const permissions = require('../../permissions');
+const socket = require('../../socket');
+
+const permissionFile = require('../../../utils/permissionfile');
+
+// eslint-disable-next-line camelcase
+const { get_location } = require('../../providers/geo/darwin');
+
+const osName = process.platform.replace('win32', 'windows').replace('darwin', 'mac');
+const permissionFunction = 'check-location-perms';
 
 let emitter;
-const logger = common.logger.prefix('request_permission');
 const done = (id, err) => {
   if (!emitter) emitter = new EventEmitter();
   emitter.emit('end', id, err);
 };
-
+/**
+ * Requests native permission on MacOS.
+ *
+ * @param {function} cb - The callback function to be executed after the permission is requested.
+ * @return {void}
+ */
+// eslint-disable-next-line consistent-return
+const requestNativePermission = (cb) => {
+  if (osName.localeCompare('mac') !== 0) return cb(new Error('Action only allowed on MacOS'));
+  // eslint-disable-next-line consistent-return
+  socket.writeMessage(permissionFunction, () => {
+    const permissionNative = permissionFile.getData('nativeLocation');
+    if (permissionNative.localeCompare('false') !== 0 && permissionNative.localeCompare('true') !== 0) {
+      try {
+        get_location((err) => {
+          cb(err);
+        });
+      } catch (ex) {
+        cb(new Error(ex.message));
+      }
+    } else {
+      cb(null);
+    }
+  });
+};
+/**
+ * Starts the process based on the provided ID and options.
+ *
+ * @param {type} id - The ID for the process
+ * @param {type} opts - The options for the process
+ * @param {type} cb - Callback function
+ * @return {type} description of return value
+ */
 // eslint-disable-next-line consistent-return
 exports.start = (id, opts, cb) => {
   // eslint-disable-next-line consistent-return
@@ -16,7 +54,7 @@ exports.start = (id, opts, cb) => {
   if (!opts.name) return done(id, new Error('Invalid permission name'));
   switch (opts.name) {
     case 'native_location':
-      permissions.requestNativeLocation((err) => {
+      requestNativePermission((err) => {
         done(id, err);
       });
       break;
@@ -26,5 +64,4 @@ exports.start = (id, opts, cb) => {
 };
 
 exports.stop = () => {
-  logger.info('inside stop');
 };
