+++ lib/agent/permissions/index.js
@@ -2,7 +2,6 @@ const osName = process.platform.replace('win32', 'windows');
 // eslint-disable-next-line import/no-dynamic-require
 const osFunctions = require(`./${osName}`);
 
-const { getLocationPermission, requestNativeLocation } = osFunctions;
+const { getLocationPermission } = osFunctions;
 
 exports.getLocationPermission = getLocationPermission;
-exports.requestNativeLocation = requestNativeLocation;
