+++ lib/agent/permissions/linux.js
@@ -1,10 +1,4 @@
 const getLocationPermission = () => {
 };
 
-const requestNativeLocation = (cb) => {
-  if (!cb || typeof cb !== 'function') return;
-  cb();
-};
-
 exports.getLocationPermission = getLocationPermission;
-exports.requestNativeLocation = requestNativeLocation;
