+++ lib/agent/permissions/windows.js
@@ -1,9 +1,4 @@
 const getLocationPermission = () => {
 };
-const requestNativeLocation = (cb) => {
-  if (!cb || typeof cb !== 'function') return;
-  cb();
-};
 
 exports.getLocationPermission = getLocationPermission;
-exports.requestNativeLocation = requestNativeLocation;
